Move constants from constants.py into respective modules.
Constants that can reasonably be local should be local, which fixes #81. Other changes:

- Remove fps-queue, clock.get_fps() already applies windowing
- Move `generate_complementary_color` to ship.py
- Remove unused constant EPSILON
- Give Rockets separate thrust-color
- Remove ENEMY_ACTION_WEIGHTS constant. The actions are chosen deterministically at the moment. Prior to the relativity-apocalypse, the enemies had extra actions, but some got removed, leading to determinism here. This enemy-ai will probably be entirely replaced by MarkovAI, anyway.
